[Refactor] 웹서버 구조 변경
1. Server와 ServerRun 객체 통합
2. ServerModule 객체 삭제

resolved #21
