+++ Makefile
@@ -5,8 +5,6 @@ NAME = webserv
 
 SRCS =	webserve/main.cpp \
 		webserve/Server/Server.cpp \
-		webserve/ServerModule/ServerModule.cpp \
-		webserve/ServerRun/ServerRun.cpp \
 		webserve/ServerParser/ServerParser.cpp \
 		webserve/ServerHandler/ServerHandler.cpp \
 		webserve/WebServer/WebServer.cpp \
@@ -19,9 +17,7 @@ SRCS =	webserve/main.cpp \
 
 OBJS = $(SRCS:.cpp=.o)
 HEAD =  -I./webserve/Server \
-		-I./webserve/ServerModule \
 		-I./webserve/ServerParser \
-		-I./webserve/ServerRun \
 		-I./webserve/ServerHandler \
 		-I./WebServer \
 		-I./webserve/HttpMessage \
