+++ webserve/ServerHandler/ServerHandler.hpp
@@ -4,15 +4,15 @@
 #include "../lib/ft/ft.hpp"
 #include "../HttpMessage/HttpRequestMessage.hpp"
 #include "../HttpMessage/HttpResponseMessage.hpp"
-#include "../ServerModule/ServerModule.hpp"
+#include "../ServerParser/ServerParser.hpp"
 #include <map>
 
-class Server;
-class ServerHandler : public ServerModule
+class ServerHandler
 {
     private:
-    std::map<int, std::string> _status;
-    HttpRequestMessage _request_message;
+    std::map<int, std::string>  _status;
+    HttpRequestMessage          _request_message;
+    ServerParser                _server_parser;
 
     void init_status();
     bool checkFile(std::string request_target);
@@ -26,7 +26,7 @@ class ServerHandler : public ServerModule
 
     public:
     ServerHandler();
-    ServerHandler(Server* self);
+    ServerHandler(ServerParser server_parser);
     ServerHandler(const ServerHandler& server_handler);
     ~ServerHandler();
     ServerHandler& operator=(const ServerHandler& server_handler);
